Ship weapon level and simple hit box
I have an interface that should be implemented by the ship class to
store weapon level. I set up some simple hit boxes for the mock ship and
enemy I have to begin testing the weapon upgrades.

Also, the debugger is fucked right now and I think it might be from the
package issues with the files. I wasted so much time trying to use the
broken debugger and had to create a new project on my client to debug
with everything in the same package. Does anyone know how to fix this
issue?